django_nginx_proxy/requirements/base.txt
Django==1.10.4

# Configuration
django-environ==0.4.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Images
Pillow==4.1.1

# DRFs
djangorestframework==3.6.2
